코드 정리 및 간소화, 예외처리 진행

+++ dailymoon/src/main/java/com/example/dailymoon/service/DiaryServiceImpl.java
@@ -3,7 +3,6 @@ package com.example.dailymoon.service;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.NoSuchElementException;
 
 import javax.transaction.Transactional;
 
@@ -15,11 +14,13 @@ import com.amazonaws.AmazonServiceException;
 import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
+import com.example.dailymoon.common.ErrorCode;
+import com.example.dailymoon.common.exception.ApiControllerException;
+import com.example.dailymoon.dto.CalanderDTO;
 import com.example.dailymoon.dto.DiaryDTO;
 import com.example.dailymoon.dto.FileDTO;
 import com.example.dailymoon.entity.Diary;
 import com.example.dailymoon.entity.Member;
-import com.example.dailymoon.form.PreviewDiary;
 import com.example.dailymoon.repository.DiaryRepository;
 import com.example.dailymoon.repository.FileRepository;
 import com.example.dailymoon.repository.MemberRepository;
@@ -29,12 +30,11 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class DiaryServiceImpl implements DiaryService {
+	
 	@Autowired
 	DiaryRepository diaryRepo;
 	@Autowired
 	MemberRepository memberRepo;
-	@Autowired
-	FileServiceImpl fileService;
 	@Autowired
 	FileRepository fileRepo;
 	
@@ -48,22 +48,21 @@ public class DiaryServiceImpl implements DiaryService {
 	// [Create]==============================================================================================================================
 	@Override
 	@Transactional
-	public void createDiary(Long userId, DiaryDTO diaryDTO, LocalDate date) {
+	public Diary createDiary(Long userId, CalanderDTO diaryDTO, LocalDate date) {
 		// member객체, date 삽입
 		Member member = memberRepo.findByUserId(userId);
-
-		if (!diaryRepo.ExistByMemberAndDate(member, date)) {
-			// diary 객체 생성
-			Diary diary = DiaryDTO.diaryDTOToEntity(diaryDTO);
-
-			diary.memberInDiary(member);
-			diary.dateInDiary(date);
-
+		
+		Diary diaryEntity = diaryRepo.findByMemberAndDate(member, date).orElseGet(Diary::new);
+		if (diaryEntity.getDiaryNo() != null) {
+			throw new ApiControllerException(ErrorCode.CONFILICT);
+		} else {
+			diaryEntity = Diary.builder().date(date).member(member).detail(diaryDTO.getDetail())
+					.feeling(diaryDTO.getFeeling()).build();
 			// DB삽입
-			diaryRepo.save(diary);
+			diaryRepo.save(diaryEntity);
+	
 		}
-		
-		
+		return diaryEntity;
 	}
 	
 
@@ -71,46 +70,50 @@ public class DiaryServiceImpl implements DiaryService {
 	@Override
 	@Transactional
 	public DiaryDTO loadDiaryDTO(Long userId, LocalDate date) {
-		Diary diary = diaryRepo.findByUserIdAndDate(userId, date);
-		DiaryDTO diaryDTO = Diary.diaryEntityToDTO(diary);
-		return diaryDTO;
+		
+		Diary diary = diaryRepo.findByUserIdAndDate(userId, date).orElseThrow(() -> new ApiControllerException(ErrorCode.POSTS_NOT_FOUND));
+
+		return DiaryDTO.diaryEntityToDTO(diary);
 	}
 	
 	// 해당 달의 다이어리 가져오기
 	@Override
 	@Transactional
-	public List<PreviewDiary> loadAllDairyDTO(Long userId) {
+	public List<CalanderDTO> loadCalenderDTO(Long userId) {
 		List<Diary> allDiaryEntity = diaryRepo.findByUserId(userId);
 
-		List<PreviewDiary> allDiary = new ArrayList<PreviewDiary>();
+		List<CalanderDTO> calanderDTO = new ArrayList<CalanderDTO>();
 		for(Diary diary:allDiaryEntity) {
-			PreviewDiary previewDiary = new PreviewDiary();
-			previewDiary.setDate(diary.getDate().toString());
-			previewDiary.setFeeling(diary.getFeeling());
-			allDiary.add(previewDiary);
+			calanderDTO.add(CalanderDTO.diaryEntityToDTO(diary));
 		}
-		return allDiary;
+		return calanderDTO;
 	}
 	
 	// [Update]===============================================================================================================================
 	@Override
 	@Transactional
-	public DiaryDTO updateDiary(DiaryDTO diaryDTO) {
-		Diary diary = diaryRepo.findById(diaryDTO.getDiaryNo()).orElseThrow(NoSuchElementException::new);
+	public DiaryDTO updateDiary(Long userId,DiaryDTO diaryDTO) {
+		
+		Diary diary = diaryRepo.findById(diaryDTO.getDiaryNo()).orElseThrow(() -> new ApiControllerException(ErrorCode.POSTS_NOT_FOUND));
+		if(!diary.getMember().getUserId().equals(userId)) {
+			throw new ApiControllerException(ErrorCode.FORBIDDEN);
+		}
 		diary.updateDiary(diaryDTO);
-		return Diary.diaryEntityToDTO(diary);
+		return DiaryDTO.diaryEntityToDTO(diary);
 	}
 	
 	// [Delete]===============================================================================================================================
 	@Override
 	@Transactional
-	public void deleteDiary(DiaryDTO diaryDTO) {
-		
-		
+	public void deleteDiary(Long userId, DiaryDTO diaryDTO) {
+		Diary diary = diaryRepo.findById(diaryDTO.getDiaryNo()).orElseThrow(()-> new ApiControllerException(ErrorCode.POSTS_NOT_FOUND));
+		if(!diary.getMember().getUserId().equals(userId)) {
+			throw new ApiControllerException(ErrorCode.FORBIDDEN);
+		}
 		try {
 			for(FileDTO file:diaryDTO.getFiles()) {
 				amazonS3Client.deleteObject(new DeleteObjectRequest(S3Bucket, file.getFileName()));
-				
+				System.out.println(file.getFileName());
 			}
 		} catch(AmazonServiceException e) {
 			e.printStackTrace();
